Program to convert text json to training data for fine-tuning an LLM
This script is used to generate the dataset files for LLM fine-tuning.
The fine-tuning goal is to parse non-standard user description into a standard JSON format.
Text files are used as the input, and the json files are used as the response.
Output is written to the dataset directory.
Usage:
go run dataset.go
Before running, make sure to place the text files and json files in the same directory as this script.
The text files should contain the user descriptions, and the json files should contain the expected responses.
The json files should have the same name as the txt files, but with the .json extension.
Example:
text file: 1.txt
json file: 1.json